requirements.txt
Flask==1.1.1
Werkzeug==0.16.0
Jinja2==2.11.1
MarkupSafe==1.1.1
itsdangerous==1.1.0
Click==7.1.1
asyncio>=3.4.3
requests==2.23.0
botbuilder-core==4.13.0
botframework-connector==4.13.0
botbuilder-schema==4.13.0
botbuilder-dialogs==4.13.0
botbuilder-ai==4.13.0
botbuilder-applicationinsights==4.13.0
botbuilder-integration-applicationinsights-aiohttp==4.13.0
azure-cognitiveservices-language-luis==0.2.0
datatypes-date-time>=1.0.0.a1
msrest==0.6.10
unittest2
aiounittest
opencensus==0.11.1
opencensus-context==0.1.3
opencensus-ext-azure==1.1.8
emoji==1.7.0